+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1263,5 +1263,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-02T08:51:22.358Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-02T10:57:37.591Z")
     r"""The date and time the resource was last updated"""