+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-02T08:51:25.702Z")
+    created: Optional[datetime] = parse_datetime("2025-09-02T10:57:41.007Z")
     r"""The date and time the resource was created"""