+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K44SZJCHK841WF7TPGKFR4PE"
+        "01K4516RCAJ98EYMBG0S73JGPV"
     )
     r"""The id of the resource"""