+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K44SZJCKG2YPD5716TT10C5Q"
+        "01K4516RCCS53BNCPM7EYME6Z5"
     )
     r"""The id of the resource"""